requirements.txt
Flask>=3.0,<4.0
numpy>=1.24
tensorflow>=2.14
nltk>=3.8
fpdf2>=2.7
